refactor(form-fields): split label stack helpers
Keep the form-field stack facade stable while moving prompt continuation and stacked-label collection into focused modules. This makes future label heuristics easier to review without changing the expansion entry points.
